+++ src/atoms/cart.ts
@@ -1,7 +1,6 @@
 import { selector, atom, selectorFamily } from 'recoil';
 import { CartItem } from '../types/cart';
 import { fetchCart } from '../apis/cart';
-import { ENDPOINT } from '../constants/auth';
 
 export const cartState = atom({
   key: 'cart',
@@ -82,9 +81,4 @@ export const totalPriceSelector = selector({
       0
     );
   },
-});
-
-export const endpointKeyState = atom<keyof typeof ENDPOINT>({
-  key: 'endpointKeyState',
-  default: '말랑',
 });
