+++ src/atoms/endpoint.ts
@@ -1,0 +1,7 @@
+import { atom } from 'recoil';
+import { ENDPOINT } from '../constants/auth';
+
+export const endpointKeyState = atom<keyof typeof ENDPOINT>({
+  key: 'endpointKeyState',
+  default: '말랑',
+});
